+++ src/types/messages.ts
@@ -20,7 +20,6 @@ export type ToAgentMessage =
   | RootDestroyWidgetMessage
   | RootResetWidgetMessage
   | RootTriggerWidgetMessage
-  | RootSetGuardContextMessage
   | RootStoreSetReplyMessage
   | RootStoreGetReplyMessage
   | RootSignalsGetReplyMessage
@@ -107,11 +106,6 @@ export interface RootResetWidgetMessage {
 export interface RootDestroyWidgetMessage {
   type: "root_destroy_widget";
 }
-
-export interface RootSetGuardContextMessage {
-  type: "root_set_guard_context";
-  guard_context: string;
-}
 // Messages to the widget from the agent
 
 export interface WidgetSetStateMessage {
@@ -189,7 +183,6 @@ export interface RootRiskIntelligenceGenerateMessage {
   type: "root_risk_intelligence_generate";
   uid: string;
   sitekey: string;
-  guard_context?: string;
   bypassCache: boolean;
 }